Make cmd testable
